updated plugin
