+++ .github/e2e-test/notation-plugin/LICENSE
