+++ .github/e2e-test/notation-plugin/Makefile
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-COMMANDS       = notation-local-signer
+COMMANDS       = e2e-test-notation-plugin
 LDFLAGS        =
 GO_BUILD_FLAGS = --ldflags="$(LDFLAGS)"
