+++ .github/e2e-test/notation-plugin/README.md
@@ -1,0 +1,2 @@
+# e2e-test-notation-plugin
+e2e-test-notation-plugin is for notation github actions E2E test workflows.
