--- .github/notation-plugin/README.md
@@ -1,2 +1,0 @@
-# Notation Local Signer plugin for Github Action
-notation-local-signer plugin is for notation github actions E2E test workflows.
